refactor(serialization): add word serialization module path
