+++ basic_system/src/system_implementation/flat_storage_model/account_cache_entry.rs
@@ -453,7 +453,7 @@ mod tests {
     use storage_models::common_structs::PreimageCacheModel;
     use zk_ee::common_structs::PreimageType;
     use zk_ee::execution_environment_type::ExecutionEnvironmentType;
-    use zk_ee::oracle::usize_serialization::{WordDeserializable, WordSerializable};
+    use zk_ee::oracle::word_serialization::{WordDeserializable, WordSerializable};
     use zk_ee::oracle::IOOracle;
     use zk_ee::reference_implementations::{BaseResources, DecreasingNative};
     use zk_ee::system::errors::internal::InternalError;
