+++ callable_oracles/src/lib.rs
@@ -22,7 +22,7 @@ pub mod blob_kzg_commitment;
 pub mod utils;
 
 use zk_ee::{
-    oracle::usize_serialization::{WordDeserializable, WordSerializable, WordSink},
+    oracle::word_serialization::{WordDeserializable, WordSerializable, WordSink},
     system::errors::internal::InternalError,
 };
