+++ forward_system/src/run/query_processors/mod.rs
@@ -1,7 +1,7 @@
 use oracle_provider::MemorySource;
 use oracle_provider::OracleQueryProcessor;
 use serde::{Deserialize, Serialize};
-use zk_ee::oracle::usize_serialization::dyn_usize_iterator::DynWordIterator;
+use zk_ee::oracle::word_serialization::dyn_word_iterator::DynWordIterator;
 
 // Oracle query processors for the forward running system.
 // Each processor handles specific types of oracle queries.
