+++ forward_system/src/run/query_processors/tx_data.rs
@@ -8,7 +8,7 @@ use zk_ee::oracle::query_ids::TX_FROM_QUERY_ID;
 use zk_ee::oracle::query_ids::{
     NEXT_TX_SIZE_QUERY_ID, TX_DATA_WORDS_QUERY_ID, TX_ENCODING_FORMAT_QUERY_ID,
 };
-use zk_ee::oracle::usize_serialization::dyn_usize_iterator::DynWordIterator;
+use zk_ee::oracle::word_serialization::dyn_word_iterator::DynWordIterator;
 use zk_ee::utils::usize_rw::ReadIterWrapper;
 
 /// This processor handles four types of queries:
