+++ oracle_provider/src/lib.rs
@@ -8,7 +8,7 @@ use std::cell::RefCell;
 use std::collections::BTreeMap;
 use std::rc::Rc;
 use zk_ee::oracle::query_ids::{DISCONNECT_ORACLE_QUERY_ID, UART_QUERY_ID};
-use zk_ee::oracle::usize_serialization::{WordDeserializable, WordSerializable};
+use zk_ee::oracle::word_serialization::{WordDeserializable, WordSerializable};
 use zk_ee::system::errors::internal::InternalError;
 use zk_ee::{internal_error, oracle::IOOracle};
