+++ zk_ee/src/execution_environment_type.rs
@@ -1,5 +1,5 @@
 use crate::internal_error;
-use crate::oracle::usize_serialization::{WordDeserializable, WordSerializable, WordSink};
+use crate::oracle::word_serialization::{WordDeserializable, WordSerializable, WordSink};
 use crate::system::errors::internal::InternalError;
 
 #[repr(u8)]
