+++ zk_ee/src/oracle/word_serialization/mod.rs
@@ -1,0 +1,5 @@
+pub use super::usize_serialization::*;
+
+pub mod dyn_word_iterator {
+    pub use super::super::usize_serialization::dyn_usize_iterator::*;
+}
